requirements.txt
Flask==2.2.2
gunicorn
Werkzeug==2.2.2
flask-cors
pymupdf
# CRITICAL: Pin exact versions to fix Azure App Service compatibility
typing_extensions==4.15.0
# Pin pydantic versions for compatibility
pydantic==2.12.4
pydantic-core==2.41.5
langchain>=0.1.0
langchain_community>=0.0.10
langchain_openai>=0.0.5
langchain_chroma>=0.1.0
langchain-core>=0.1.0
pypdf
cryptography
transformers
PyMuPDF

# Load env vars from .env
python-dotenv>=1.0.0

# Use bundled SQLite on Linux for Chroma persistence
pysqlite3-binary; platform_system == "Linux"

paperscraper
chromadb>=0.4.24
openai>=1.0.0

# Pin OpenTelemetry packages to match ChromaDB requirements and override system defaults
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp-proto-grpc>=1.27.0
opentelemetry-instrumentation>=0.48b0